pyproject.toml
[project]
name = "danielv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateparser>=1.2.1",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.15.0",
    "hrequests[all]>=0.9.2",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "json-repair>=0.44.1",
    "ptpython>=3.0.30",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "the-retry>=0.1.1",
]
